lite/examples/recommendation/ml/requirements.txt
pandas>=1.0.5
tensorflow>=2.2.0
absl-py>=0.1.6
